--- src/Cook/Pages/Cook_Register/cook_register.css
@@ -1,43 +1,0 @@
-.register{
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-form{
-    display: flex;
-    /* justify-content: center; 
-     align-items: center; */
-    flex-direction: column;
-    
-}
-
-.register-heading{
-    font-family: 'Courier New', Courier, monospace;
-    font-size: 2em;
-    margin-top: 30px;
-    text-align: center;
-}
-
-input{
-    margin-top: 10px;
-    border-radius: 5px;
-    height: 30px;
-}
-
-.reg-button{
-    margin-top: 10px;
-    border-radius: 5px;
-    height: 30px;
-    background-color: gray;
-    color: white;
-    
-}
-
-.register-login{
-    text-align: center;
-    margin-top: 10px;
-}
