+++ src/Customer/pages/Customer_Login/customer_login.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useState } from 'react'
 import { Link } from 'react-router-dom'
-import '../Login/login.css'
+import './customer_login.css'
 
 const Login = () => {
